# backend/docs/social-api-draft.openapi.yaml
...
openapi: 3.0.3
info:
  title: STROITECH Social API (draft)
  version: "0.1.0"
  description: >
    Планируемые маршруты. Префикс API в проде обычно /api (как у текущих /api/users, /api/feed).
servers:
  - url: /api
...
paths:
...
  /communities:
    get:
      tags: [communities]
      summary: Список сообществ
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: region
          in: query
          schema: { type: string }
        - $ref: "#/components/parameters/page"
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
      responses:
        "200":
          description: OK
    post:
      tags: [communities]
      summary: Создать сообщество (создатель — admin)
      security: [{ bearerAuth: [] }]
      responses:
        "201":
          description: OK
  /communities/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [communities]
      summary: Карточка сообщества + myRole при JWT
      responses:
        "200":
          description: OK
    patch:
      tags: [communities]
      summary: Редактирование (admin сообщества или глобальный moderator)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
    delete:
      tags: [communities]
      summary: Удалить (создатель или глобальный moderator)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
  /communities/{id}/members:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [communities]
      summary: Участники с ролями
      parameters:
        - $ref: "#/components/parameters/page"
        - name: limit
          in: query
          schema: { type: integer, default: 30, maximum: 100 }
      responses:
        "200":
          description: OK
  /communities/{id}/albums:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [communities, albums]
      summary: Альбомы, привязанные к сообществу
      responses:
        "200":
          description: OK
  /communities/{id}/join:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [communities]
      summary: Вступить (открытая группа)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
  /communities/{id}/leave:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [communities]
      summary: Выйти (единственный admin — с ограничениями)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
  /communities/{id}/members/{userId}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
      - name: userId
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      tags: [communities]
      summary: Исключить участника (staff сообщества или глобальный moderator)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
    patch:
      tags: [communities]
      summary: Сменить роль moderator|member (только admin сообщества)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
...